ragnarok-install: rm do_install func
This whole script needs to be simplified. There is 0 upside to asking
question before doing the actually install, but it does increase the
size and complexity of the install script, so go back to the simpler way
of "ask question -> configure the system".

Installing from scratch with mmdebstrap will use mkconf, but first, the
focus should be on the basic installer (which fetches base01.tgz from
the net).